+++ SQL/postgres/2022081200.sql
@@ -1,0 +1,2 @@
+DROP INDEX responses_user_id_idx;
+CREATE INDEX responses_user_id_idx ON responses (user_id, del);