+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix bug where renamed options were removed from the config on installto.sh (update.sh) run (#8643)
 - Fix favicon rewrite rule in .htaccess (#8654)
 - Fix various PHP 8.2 warnings
+- Fix bug where it wasn't possible to create more than one response record on SQLite and Postgres (#8664)
 
 ## Release 1.6.0